+++ R/HttpLibAdapterRegistry.R
@@ -34,10 +34,10 @@ HttpLibAdapaterRegistry <- R6::R6Class(
 
     register = function(x) {
       # FIXME: when other adapters supported, change this inherits test
-      if (!inherits(x, "CrulAdapter")) {
+      if (!inherits(x, c("CrulAdapter", "HttrAdapter", "CurlAdapter"))) {
         stop("'x' must be an adapter, such as CrulAdapter", call. = FALSE)
       }
-      self$adapters <- cc(list(self$adapters, x))
+      self$adapters <- c(self$adapters, x)
     }
   )
 )