+++ R/stub_request.R
@@ -34,13 +34,13 @@
 #'    wi_th(headers = list('User-Agent' = 'R'))
 #'
 #' # request body
-#' stub_request("get", "https://httpbin.org/get") %>%
+#' stub_request("post", "https://httpbin.org/post") %>%
 #'    wi_th(body = list(foo = 'bar'))
 #' stub_registry()
 #' library(crul)
 #' x <- crul::HttpClient$new(url = "https://httpbin.org")
 #' crul::mock()
-#' x$get('get')
+#' x$post('post', body = list(foo = 'bar'))
 #'
 #' # add expectation with to_return
 #' stub_request("get", "https://httpbin.org/get") %>%