+++ R/webmockr-opts.R
@@ -124,7 +124,8 @@ net_connect_explicit_allowed <- function(allowed, uri = NULL) {
 #' @export
 print.webmockr_config <- function(x, ...) {
   cat("<webmockr configuration>", sep = "\n")
-  cat(paste0("  enabled?: ", webmockr_lightswitch$crul), sep = "\n")
+  cat(paste0("  crul enabled?: ", webmockr_lightswitch$crul), sep = "\n")
+  cat(paste0("  httr enabled?: ", webmockr_lightswitch$httr), sep = "\n")
   cat(paste0("  allow_net_connect?: ", x$allow_net_connect), sep = "\n")
   cat(paste0("  allow_localhost?: ", x$allow_localhost), sep = "\n")
   cat(paste0("  allow: ", x$allow %||% ""), sep = "\n")