+++ tests/testthat/test-no-cassette-in-use.R
@@ -1,10 +1,11 @@
 context("no_cassette_in_use")
 
-library(vcr)
-dir <- tempdir()
-invisible(vcr_configure(dir = dir))
-
 test_that("no cassette in use behaves as expected", {
+  skip_if_not_installed('vcr')
+  library(vcr)
+  dir <- tempdir()
+  invisible(vcr_configure(dir = dir))
+
   crul::mock()
   x <- crul::HttpClient$new(url = "https://httpbin.org")
 
@@ -13,14 +14,13 @@ test_that("no cassette in use behaves as expected", {
     x$get("get"),
     "There is currently no cassette in use"
   )
+
+  # cleanup
+  unlink(file.path(vcr_configuration()$dir, "turtle.yml"))
+
+  # reset configuration
+  vcr_configure_reset()
+
+  # unload vcr
+  unloadNamespace("vcr")
 })
-
-
-# cleanup
-unlink(file.path(vcr_configuration()$dir, "turtle.yml"))
-
-# reset configuration
-vcr_configure_reset()
-
-# unload vcr
-unloadNamespace("vcr")